[TL-49] modified script for configuring master/slave for each port
Used the phymode in uboot env variable to configure the T1 phy mode for
each port. Each bit corresponds to one port. so lan1 corresponds to bit
1, lan2 to 2 and so on.
For example, if phymode=0x0A, it configures lan1 & lan3 as master and
lan2 & lan4 as slave.
Instead of using the ethtool for reading master/slave configuration,
	used the /sys/class/net/lan1/of_node/phy-mode. Because when
	using the ethtool, it takes 1sec for each port due to sqi. Have
	to find ways to reduce the time taken.