[TL-72] Fix for the tc policer
Earlier the tc policer command must be accompanied by the action
mirred for redirection.
Now, based on the qci result, the packets may be dropped or forwarded.
For this tcam entry is updated for the tc policer action. Now Map Mode
is set to 2 and Destion Port is set to 0xFF.
So, when qci decides the packet to be dropped, the value is 0x00 which
is and with 0xff which means packet is dropped.
If qci decides the packet is forwarded, the value is and with 0xff,
   which is port retrieved from the alu table. So the packet will be
   forwarded to the respective port without the need of action mirrored.