[TL-9] fixed the printing of ff:ff:ff:ff:ff:ff in the bridge fdb show
In the last entry, code checks the forward port & BIT(port). Since the
forward port is 0xff, the condition becomes true, so it is forwarded to
user space application. It is not expected.
So the if loop is added with one more condition, to check whether the
entry is valid or not.